requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1

# Database
psycopg2-binary==2.9.9

# Asynchronous Tasks
celery==5.3.6
redis==5.0.1

# Production Server
gunicorn==21.2.0

# Environment Variables
python-decouple==3.8

# Image Processing
Pillow==10.2.0

# API Documentation
drf-yasg==1.21.7

# Date/Time utilities
python-dateutil==2.8.2

# Security
django-cryptography==1.1

# Monitoring and Performance
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
